Added GpioManager, made some documentation changes, altered SyringePump
The GPIO Manager class is now available. You declare one of them. Then, when you need a GPIO pin, you just ask for either the next input or next output, and it will return a pin if possible.
It can also accept an MCP23017 GPIO expander chip and handle allocation of those as well.